requirements.txt
discord.py==2.1.0
aiosqlite==0.17.0
pyyaml==6.0